move all render related methods to PowerGogglesRenderer
